[NR-96048] Add VariantAdapter and class transfrom classes
Refactor to variants to use androidComponents as well as android extension.
For Gradle installations 7.0 and higher. Refactor variant plugin code in
the plugin into a controller to manage the old and new plugin variant API's.

The adapter provides agnostic methods for mutating the DAG with task
we need to instrument, configure and proces obfuscation maps.