[NR-96048] Add VariantAdapter and class transform classes
Refactor variants to use androidComponents as well as the legacy
android extension, for Gradle installations 7.2 and higher.

Refactor variant code in the plugin into an adapter to manage the old and
new plugin variant API's. The adapter provides agnostic methods for mutating
the DAG with tasks we need to instrument, configure and process obfuscation maps.