[NR-152229] Silence "duplicate entry" errors during instrumentation
This is not a bug, but rather a side-effect of AGP's instrumentation model.
Prior, JAR files were rewritten in place, and with updated model all
instrumented classes to into the same JAR. So if dependencies share the same
package names, manifest and directory entries will be rejected as
duplicates when processed. As long as these aren't class files, the error
is misleading.