[NR-170090] Cache log data on device for later delvery
This is essntially _store-and-forward_ of submitted remote log data.
The approach is a WIP, derived for the POC is to write to a single data
file, then roll (or rotate) that over when:
* a period of time has elapsed (harvest cycle, in this case)
* the uncompressed file contents exceeds the Vortex ingest limit
* the agent is shut down

On harvest, the logs report directory is swept to:
* roll all closed log data files into a common archive (1 in most cases unless
  the opereration failed in previous attempts)
* forward all available archives to log ingest
* remove any data that's exceeded its use-by date

Submitted log data is async-safe and uses a thread pool to handle incoming requests.
Rolling log files as data continues to arrive is a critical section.
Log forwarding is disabled in this commit.

Log data files are not explicitly delete, but rather renamed as backups and
expired later. This provides some recovery for now, but will likely
be changed to GA.

All this shoudl be reflected in unit tests.

TODO:
* move threading concerns from RemoteLogger to reporter (forwarder)
* add a log data validator to enforce Log ingest constraints, anonymize confidential data
* add a filesystem storage budget and policy to prevent consuming all device storage