[NR-230991] Record ApplicationExitInfo for Android 11+
The agent sweeps for data for app exit info once, during initialization.
ApplicationEventInfo records are stored as files in app's cache dir,
which also serves to indicate an app exit info entry has been recorded.

Exit info is stored in a new (to be) reserved event type and deliverd with
the first harvest.

* requires upgrade to `compileSdk 30` to support new API
* creates new event category: ApplicationExit
* creates new event type: MobileApplicationExit
* creates new attribute set for event
* adds tests; refactors test code to share stubs